♻️ refactor(workers): add constructor DI to JobsWorker and add tests
- inject optional JobDao via constructor for testability
- use $this->jobDao instead of new JobDao() in _recountAvgPee()
- add tests for process(), requeue threshold, and PEE computation
- achieve 100% coverage on JobsWorker

+++ lib/Model/Jobs/JobStruct.php
@@ -55,7 +55,7 @@ class JobStruct extends AbstractDaoSilentStruct implements IDaoStruct, ArrayAcce
     public ?string $id_translator = null;
     public ?string $job_type = null;
     public int $total_time_to_edit = 0;
-    public int $avg_post_editing_effort = 0;
+    public float $avg_post_editing_effort = 0.0;
     public int $only_private_tm = 0;
     public ?int $last_opened_segment = null;
     public int $id_tms = 1;
